Fixed text color calculation

+++ src/main/java/dreampool/ui/parts/Text.java
@@ -1,6 +1,5 @@
 package dreampool.ui.parts;
 
-import org.joml.Matrix4f;
 import org.joml.Vector2f;
 import org.joml.Vector3f;
 import org.joml.Vector4f;
@@ -20,7 +19,7 @@ public class Text extends Part{
 	public Text(String text, float size, float x, float y, int r, int g, int b, float a, Font font) {
 		this.text = text;
 		position = new Vector2f(x / Application.resDivisor, y / Application.resDivisor);
-		color = new Vector4f(r / 255, g / 255, b / 255, a);
+		color = new Vector4f((float)r / 255, (float)g / 255, (float)b / 255, a);
 		this.font = font;
 		this.size = size * Application.resDivisor;
 	}
